Read bills from the frontend
